refactor(aether-actor): act on the audit-quality panel's cleanup findings
Apply the crate-internal cleanups the /audit-quality panel surfaced on
aether-actor (#2436): control-flow collapses, the guest_alloc module
move and the emit_log_event cohesion split, a ChainMode enum over the
route_or_enqueue bool, the WasmCtx::sender retype to Option<ReplyHandle>,
two in-crate-only visibility narrowings, and name/import tidies. The
breaking public-API harmonizations are routed to the issue's side
findings, not applied here.
