+++ crates/aether-actor/src/wasm/bridge/mod.rs
@@ -8,11 +8,12 @@
 //! audited ptr/len marshalling) while closing the over-exposure the
 //! `pub static` forms created.
 //!
+//! - `log` — log-event FFI (`emit_log_event`). Split from `mail` because
+//!   it is a distinct op family with no relation to mail routing.
 //! - `mail` — outbound mail (`send_mail`, `reply_mail`,
-//!   `prev_correlation`, `emit_log_event`, `spawn_sibling`,
-//!   `spawn_inline_child`). Correlation lives here because every send
-//!   mints one so a handler can match a reply to the request it sent —
-//!   it's mail-level metadata.
+//!   `prev_correlation`, `spawn_sibling`, `spawn_inline_child`).
+//!   Correlation lives here because every send mints one so a handler
+//!   can match a reply to the request it sent — it's mail-level metadata.
 //! - `persist` — migration-bundle deposit
 //!   (`save_state`), used during `on_dehydrate` only.
 //!
@@ -21,5 +22,6 @@
 //! per-stage capability traits in [`crate::model::ctx`], not a single
 //! transport trait.
 
+pub(crate) mod log;
 pub(crate) mod mail;
 pub(crate) mod persist;
